+++ Revo.Core/Events/EventMessageDraft.cs
@@ -1,15 +1,35 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Revo.Core.Events
 {
+    public static class EventMessageDraft
+    {
+        public static IEventMessageDraft<TEvent> FromEvent<TEvent>(TEvent @event)
+            where TEvent : IEvent
+        {
+            Type messageType = typeof(EventMessageDraft<>).MakeGenericType(@event.GetType());
+            return (IEventMessageDraft<TEvent>)messageType.GetConstructor(new[] {@event.GetType()})
+                .Invoke(new object[] { @event });
+        }
+    }
+
     public class EventMessageDraft<TEvent> : IEventMessageDraft<TEvent>
         where TEvent : IEvent
     {
-        private readonly Dictionary<string, string> metadata = new Dictionary<string, string>();
+        private readonly Dictionary<string, string> metadata;
 
         public EventMessageDraft(TEvent @event)
         {
             Event = @event;
+            metadata = new Dictionary<string, string>();
+        }
+        
+        public EventMessageDraft(TEvent @event, IReadOnlyDictionary<string, string> metadata)
+        {
+            Event = @event;
+            this.metadata = metadata.ToDictionary(x => x.Key, x => x.Value);
         }
 
         IEvent IEventMessage.Event => Event;
@@ -17,9 +37,10 @@ namespace Revo.Core.Events
         public TEvent Event { get; }
         public IReadOnlyDictionary<string, string> Metadata => metadata;
 
-        public void SetMetadata(string key, string value)
+        public IEventMessageDraft SetMetadata(string key, string value)
         {
             metadata[key] = value;
+            return this;
         }
     }
 }