+++ Revo.Infrastructure/EventStores/Generic/EventStoreRecordAdapter.cs
@@ -7,7 +7,7 @@ using Revo.Infrastructure.EventStores.Generic.Model;
 
 namespace Revo.Infrastructure.EventStores.Generic
 {
-    public class EventStoreRecordAdapter : IEventStoreRecord
+    public class EventStoreRecordAdapter : IEventStreamRowEventStoreRecord
     {
         private readonly EventStreamRow row;
         private readonly IEventSerializer eventSerializer;
@@ -20,12 +20,10 @@ namespace Revo.Infrastructure.EventStores.Generic
             this.eventSerializer = eventSerializer;
         }
 
-        public IEvent Event => @event ??
-                               (@event = eventSerializer.DeserializeEvent(row.EventJson,
-                                   new VersionedTypeId(row.EventName, row.EventVersion)));
+        public IEvent Event => @event ??= eventSerializer.DeserializeEvent(row.EventJson,
+            new VersionedTypeId(row.EventName, row.EventVersion));
 
-        public IReadOnlyDictionary<string, string> AdditionalMetadata => additionalMetadata
-                                                                         ?? (additionalMetadata = eventSerializer.DeserializeEventMetadata(row.AdditionalMetadataJson));
+        public IReadOnlyDictionary<string, string> AdditionalMetadata => additionalMetadata ??= eventSerializer.DeserializeEventMetadata(row.AdditionalMetadataJson);
 
         public Guid EventId => row.Id;
         public long StreamSequenceNumber => row.StreamSequenceNumber;