+++ Revo.Infrastructure/EventStores/Generic/IEventStreamRowEventStoreRecord.cs
@@ -1,0 +1,9 @@
+﻿using Revo.Infrastructure.EventStores.Generic.Model;
+
+namespace Revo.Infrastructure.EventStores.Generic
+{
+    public interface IEventStreamRowEventStoreRecord : IEventStoreRecord
+    {
+        EventStreamRow EventStreamRow { get; }
+    }
+}