+++ Revo.Infrastructure/Events/Async/Generic/AsyncEventQueueManager.cs
@@ -14,16 +14,17 @@ namespace Revo.Infrastructure.Events.Async.Generic
     public class AsyncEventQueueManager : IAsyncEventQueueManager
     {
         private readonly ICrudRepository crudRepository;
-        private readonly IEventSerializer eventSerializer;
+        private readonly IQueuedAsyncEventMessageFactory queuedAsyncEventMessageFactory;
         private readonly IExternalEventStore externalEventStore;
 
-        private readonly Dictionary<IEventMessage, List<EventSequencing>> eventsToQueues = new Dictionary<IEventMessage, List<EventSequencing>>();
+        private readonly Dictionary<IEventMessage, List<EventSequencing>> eventsToQueues
+            = new Dictionary<IEventMessage, List<EventSequencing>>();
 
-        public AsyncEventQueueManager(ICrudRepository crudRepository, IEventSerializer eventSerializer,
+        public AsyncEventQueueManager(ICrudRepository crudRepository, IQueuedAsyncEventMessageFactory queuedAsyncEventMessageFactory,
             IExternalEventStore externalEventStore)
         {
             this.crudRepository = crudRepository;
-            this.eventSerializer = eventSerializer;
+            this.queuedAsyncEventMessageFactory = queuedAsyncEventMessageFactory;
             this.externalEventStore = externalEventStore;
         }
 
@@ -31,7 +32,7 @@ namespace Revo.Infrastructure.Events.Async.Generic
         {
             return (await QueryQueuedEvents()
                 .Where(x => asyncEventQueueRecordIds.Contains(x.Id))
-                .ToListAsync(crudRepository))
+                .ToArrayAsync(crudRepository))
                 .Select(SelectRecordFromQueuedEvent)
                 .ToList();
         }
@@ -46,13 +47,13 @@ namespace Revo.Infrastructure.Events.Async.Generic
 
         public async Task<IAsyncEventQueueState> GetQueueStateAsync(string queueName)
         {
-            AsyncEventQueue queue = await GetQueueAsync(queueName);
+            var queue = await FindQueueAsync(queueName);
             return queue; //can be null
         }
 
         public async Task<IReadOnlyCollection<IAsyncEventQueueRecord>> GetQueueEventsAsync(string queueName)
         {
-            AsyncEventQueue queue = await GetQueueAsync(queueName);
+            var queue = await FindQueueAsync(queueName);
             if (queue == null)
             {
                 return new List<IAsyncEventQueueRecord>();
@@ -82,26 +83,32 @@ namespace Revo.Infrastructure.Events.Async.Generic
             }
         }
         
-        public Task EnqueueEventAsync(IEventMessage eventMessage, IEnumerable<EventSequencing> queues)
+        public Task EnqueueEventAsync(IEventMessage eventMessage, IReadOnlyCollection<EventSequencing> queues)
         {
-            if (!eventsToQueues.TryGetValue(eventMessage, out var queueList))
+            if (!eventsToQueues.TryGetValue(eventMessage, out var queueList)
+                || (queueList == null && queues.Count > 0))
             {
-                queueList = eventsToQueues[eventMessage] = new List<EventSequencing>();
+                queueList = eventsToQueues[eventMessage] = queues.Count > 0 ? new List<EventSequencing>() : null;
             }
 
-            queueList.AddRange(queues);
+            if (queues.Count > 0)
+            {
+                // ReSharper disable once PossibleNullReferenceException
+                queueList.AddRange(queues);
+            }
+
             return Task.CompletedTask;
         }
-        
+
         public async Task<string> GetEventSourceCheckpointAsync(string eventSourceName)
         {
-            EventSourceState eventSourceState = await crudRepository.FindAsync<EventSourceState>(eventSourceName);
+            var eventSourceState = await crudRepository.FindAsync<EventSourceState>(eventSourceName);
             return eventSourceState?.EventEnqueueCheckpoint;
         }
 
         public async Task SetEventSourceCheckpointAsync(string eventSourceName, string opaqueCheckpoint)
         {
-            EventSourceState eventSourceState = await crudRepository.FindAsync<EventSourceState>(eventSourceName);
+            var eventSourceState = await crudRepository.FindAsync<EventSourceState>(eventSourceName);
             if (eventSourceState == null)
             {
                 eventSourceState = new EventSourceState(eventSourceName);
@@ -113,67 +120,48 @@ namespace Revo.Infrastructure.Events.Async.Generic
 
         public virtual async Task<IReadOnlyCollection<IAsyncEventQueueRecord>> CommitAsync()
         {
-            List<QueuedAsyncEvent> queuedEvents = await EnqueueEventsToRepositoryAsync();
+            var queuedEvents = await EnqueueEventsToRepositoryAsync();
             await crudRepository.SaveChangesAsync();
             return queuedEvents.Select(SelectRecordFromQueuedEvent).ToList();
         }
 
         protected async Task<List<QueuedAsyncEvent>> EnqueueEventsToRepositoryAsync()
         {
-            var eventStoreEvents = eventsToQueues.Where(x =>
-                (x.Key as IEventStoreEventMessage)?.Record is EventStoreRecordAdapter).ToArray();
-            var externalEvents = eventsToQueues.Except(eventStoreEvents).ToArray();
-
             var queuedEvents = new List<QueuedAsyncEvent>();
+
+            EnqueueEventStoreEvents(queuedEvents);
+            await EnqueueExternalEventsAsync(queuedEvents);
+            await CreateQueuesAsync(queuedEvents);
             
-            //external
-            if (externalEvents.Length > 0)
+            eventsToQueues.Clear();
+            return queuedEvents;
+        }
+
+        private async Task CreateQueuesAsync(List<QueuedAsyncEvent> queuedEvents)
+        {
+            var queueNames = queuedEvents.Select(x => x.QueueId).Distinct().ToArray();
+            var existingQueues = await crudRepository.FindManyAsync<AsyncEventQueue, string>(queueNames);
+            foreach (var queueName in queueNames)
             {
-                foreach (var externalEvent in externalEvents)
+                var queue = existingQueues.FirstOrDefault(x => x.Id == queueName);
+                if (queue == null)
                 {
-                    externalEventStore.PushEvent(externalEvent.Key);
-                }
-
-                var externalEventRecords = await externalEventStore.CommitAsync();
-                foreach (var externalEventRecord in externalEventRecords)
-                {
-                    if (externalEventRecord.IsDispatchedToAsyncQueues)
-                    {
-                        continue;
-                    }
-
-                    externalEventRecord.MarkDispatchedToAsyncQueues();
-                    var sequences = externalEvents.First(x => x.Key.Metadata.GetEventId() == externalEventRecord.Id).Value;
-
-                    foreach (var sequence in sequences)
-                    {
-                        AsyncEventQueue queue = await GetOrCreateQueueAsync(sequence.SequenceName, null);
-                        QueuedAsyncEvent queuedEvent = new QueuedAsyncEvent(queue.Id, externalEventRecord, sequence.EventSequenceNumber);
-                        crudRepository.Add(queuedEvent);
-                        queuedEvents.Add(queuedEvent);
-                    }
+                    queue = new AsyncEventQueue(queueName, null);
+                    crudRepository.Add(queue);
                 }
             }
+        }
 
-            //event store
-            var allQueueNames = new HashSet<string>();
+        private void EnqueueEventStoreEvents(List<QueuedAsyncEvent> queuedEvents)
+        {
+            // all event store events need to be marked as dispatched (even whey they haven't been dispatched to any queues)
+            var eventStoreEvents = eventsToQueues.Where(x =>
+                (x.Key as IEventStoreEventMessage)?.Record is IEventStreamRowEventStoreRecord).ToArray();
+
             foreach (var eventToQueues in eventStoreEvents)
             {
-                foreach (var sequence in eventToQueues.Value)
-                {
-                    allQueueNames.Add(sequence.SequenceName);
-                }
-            }
-
-            if (allQueueNames.Count > 10)
-            {
-                await LoadQueuesAsync(allQueueNames.ToArray());
-            }
-            
-            foreach (var eventToQueues in eventStoreEvents)
-            {
-                var storeMessage = (IEventStoreEventMessage) eventToQueues.Key;
-                var eventStreamRow = ((EventStoreRecordAdapter)storeMessage.Record).EventStreamRow;
+                var storeMessage = (IEventStoreEventMessage)eventToQueues.Key;
+                var eventStreamRow = ((IEventStreamRowEventStoreRecord)storeMessage.Record).EventStreamRow;
                 if (eventStreamRow.IsDispatchedToAsyncQueues)
                 {
                     continue;
@@ -187,39 +175,80 @@ namespace Revo.Infrastructure.Events.Async.Generic
                     crudRepository.SetEntityState(eventStreamRow, EntityState.Modified);
                 }
 
-                foreach (var sequence in eventToQueues.Value)
+                var eventSequencings = eventToQueues.Value;
+                if (eventSequencings != null)
                 {
-                    AsyncEventQueue queue = await GetOrCreateQueueAsync(sequence.SequenceName, null);
-                    QueuedAsyncEvent queuedEvent = new QueuedAsyncEvent(queue.Id, eventStreamRow, sequence.EventSequenceNumber);
-                    crudRepository.Add(queuedEvent);
-                    queuedEvents.Add(queuedEvent);
+                    foreach (var sequencing in eventSequencings)
+                    {
+                        var queuedEvent = new QueuedAsyncEvent(Guid.NewGuid(), sequencing.SequenceName,
+                            eventStreamRow, sequencing.EventSequenceNumber);
+                        crudRepository.Add(queuedEvent);
+                        queuedEvents.Add(queuedEvent);
+                    }
                 }
             }
-
-            eventsToQueues.Clear();
-            return queuedEvents;
         }
 
-        private Task<AsyncEventQueue> GetQueueAsync(string queueName)
+        private async Task EnqueueExternalEventsAsync(List<QueuedAsyncEvent> queuedEvents)
+        {
+            // external events are only stored if they match any queues
+            var externalEvents = eventsToQueues
+                .Where(x => !(x.Key is IEventStoreEventMessage eventStoreMessage)
+                            || !(eventStoreMessage.Record is IEventStreamRowEventStoreRecord))
+                .Where(x => x.Value?.Count > 0)
+                .Select(x =>
+                {
+                    // clone and generate event IDs where missing
+                    var eventId = x.Key.Metadata.GetEventId();
+                    if (eventId == null)
+                    {
+                        return new KeyValuePair<IEventMessage, List<EventSequencing>>(
+                            CloneMessageWithId(x.Key), x.Value);
+                    }
+
+                    return x;
+                })
+                .ToArray();
+            
+            if (externalEvents.Length > 0)
+            {
+                foreach (var externalEvent in externalEvents)
+                {
+                    externalEventStore.TryPushEvent(externalEvent.Key);
+                }
+
+                var externalEventRecords = await externalEventStore.CommitAsync();
+                foreach (var externalEventRecord in externalEventRecords)
+                {
+                    if (externalEventRecord.IsDispatchedToAsyncQueues)
+                    {
+                        continue;
+                    }
+
+                    externalEventRecord.MarkDispatchedToAsyncQueues();
+
+                    if (!crudRepository.IsAttached(externalEventRecord))
+                    {
+                        crudRepository.Attach(externalEventRecord);
+                        crudRepository.SetEntityState(externalEventRecord, EntityState.Modified);
+                    }
+
+                    var sequencings = externalEvents.First(x => x.Key.Metadata.GetEventId() == externalEventRecord.Id).Value;
+
+                    foreach (var sequencing in sequencings)
+                    {
+                        var queuedEvent = new QueuedAsyncEvent(Guid.NewGuid(), sequencing.SequenceName,
+                            externalEventRecord, sequencing.EventSequenceNumber);
+                        crudRepository.Add(queuedEvent);
+                        queuedEvents.Add(queuedEvent);
+                    }
+                }
+            }
+        }
+
+        private Task<AsyncEventQueue> FindQueueAsync(string queueName)
         {
             return crudRepository.FindAsync<AsyncEventQueue>(queueName);
-        }
-
-        private async Task LoadQueuesAsync(string[] queueNames)
-        {
-            await crudRepository.Where<AsyncEventQueue>(x => queueNames.Contains(x.Id)).ToArrayAsync(crudRepository);
-        }
-
-        private async Task<AsyncEventQueue> GetOrCreateQueueAsync(string queueName, long? lastSequenceNumberProcessed)
-        {
-            AsyncEventQueue queue = await GetQueueAsync(queueName);
-            if (queue == null)
-            {
-                queue = new AsyncEventQueue(queueName, lastSequenceNumberProcessed);
-                crudRepository.Add(queue);
-            }
-
-            return queue;
         }
         
         private IQueryable<QueuedAsyncEvent> QueryQueuedEvents()
@@ -231,9 +260,28 @@ namespace Revo.Infrastructure.Events.Async.Generic
                 .Include(crudRepository, x => x.ExternalEventRecord);
         }
 
+        private async Task<AsyncEventQueue> GetOrCreateQueueAsync(string queueName, long? lastSequenceNumberProcessed)
+        {
+            AsyncEventQueue queue = await FindQueueAsync(queueName);
+            if (queue == null)
+            {
+                queue = new AsyncEventQueue(queueName, lastSequenceNumberProcessed);
+                crudRepository.Add(queue);
+            }
+
+            return queue;
+        }
+
         protected AsyncEventQueueRecordAdapter SelectRecordFromQueuedEvent(QueuedAsyncEvent queuedEvent)
         {
-            return new AsyncEventQueueRecordAdapter(queuedEvent, eventSerializer);
+            return new AsyncEventQueueRecordAdapter(queuedEvent, queuedAsyncEventMessageFactory);
+        }
+
+        private IEventMessage CloneMessageWithId(IEventMessage eventMessage)
+        {
+            return eventMessage
+                .Clone()
+                .SetId(Guid.NewGuid());
         }
     }
 }